+++ sbi/inference/potentials/vector_field_potential.py
@@ -21,7 +21,7 @@ from sbi.neural_nets.estimators.shape_handling import (
 from sbi.samplers.ode_solvers import build_neural_ode
 from sbi.sbi_types import TorchTransform
 from sbi.utils.sbiutils import mcmc_transform, within_support
-from sbi.utils.torchutils import ensure_theta_batched
+from sbi.utils.torchutils import ensure_theta_batched, process_device
 
 
 class VectorFieldBasedPotential(BasePotential):
@@ -87,6 +87,7 @@ class VectorFieldBasedPotential(BasePotential):
             device: Device to move the score_estimator, prior and x_o to.
         """
 
+        device = process_device(device)
         self.device = device
         self.vector_field_estimator.to(device)
         if self.prior:
@@ -118,6 +119,19 @@ class VectorFieldBasedPotential(BasePotential):
                 `IIDScoreFunction`.
             ode_kwargs: Additional keyword arguments for the neural ODE.
         """
+        # IID and guidance require transforming the prior to standardized space.
+        if self.vector_field_estimator.compose_enabled:
+            if x_is_iid:
+                raise NotImplementedError(
+                    "compose_standardization does not yet support iid (x with "
+                    "batch>1). Use a single observation, or disable "
+                    "compose_standardization."
+                )
+            if guidance_method is not None:
+                raise NotImplementedError(
+                    "compose_standardization does not yet support guided sampling. "
+                    "Disable guidance, or disable compose_standardization."
+                )
         super().set_x(x_o, x_is_iid)
         self.iid_method = iid_method or self.iid_method
         self.iid_params = iid_params
@@ -155,6 +169,13 @@ class VectorFieldBasedPotential(BasePotential):
         )
         self.vector_field_estimator.eval()
 
+        compose = self.vector_field_estimator.compose_enabled
+        if compose:
+            flow_input = self.vector_field_estimator.to_z(theta_density_estimator)
+            log_abs_det = self.vector_field_estimator.log_abs_det()
+        else:
+            flow_input = theta_density_estimator
+
         with torch.set_grad_enabled(track_gradients):
             if self.x_is_iid:
                 assert self.prior is not None, (
@@ -166,10 +187,7 @@ class VectorFieldBasedPotential(BasePotential):
                 num_iid = self.x_o.shape[0]  # number of iid samples
                 iid_posteriors_prob = torch.sum(
                     torch.stack(
-                        [
-                            flow.log_prob(theta_density_estimator).squeeze(-1)
-                            for flow in self.flows
-                        ],
+                        [flow.log_prob(flow_input).squeeze(-1) for flow in self.flows],
                         dim=0,
                     ),
                     dim=0,
@@ -180,7 +198,9 @@ class VectorFieldBasedPotential(BasePotential):
                     theta_density_estimator
                 ).squeeze(-1)
             else:
-                log_probs = self.flow.log_prob(theta_density_estimator).squeeze(-1)
+                log_probs = self.flow.log_prob(flow_input).squeeze(-1)
+                if compose:
+                    log_probs = log_probs - log_abs_det
             # Force probability to be zero outside prior support.
             in_prior_support = within_support(self.prior, theta)