+++ sbi/utils/user_input_checks.py
@@ -13,7 +13,12 @@ from torch.distributions import Distribution, Uniform
 
 from sbi.sbi_types import Array
 from sbi.utils.sbiutils import within_support
-from sbi.utils.torchutils import BoxUniform, assert_all_finite, atleast_2d
+from sbi.utils.torchutils import (
+    BoxUniform,
+    assert_all_finite,
+    atleast_2d,
+    canonical_device,
+)
 from sbi.utils.user_input_checks_utils import (
     CustomPriorWrapper,
     MultipleIndependent,
@@ -758,7 +763,7 @@ def validate_theta_and_x(
     assert theta.dtype == float32, "Type of parameters must be float32."
     assert x.dtype == float32, "Type of simulator outputs must be float32."
 
-    if str(x.device) != str(data_device):
+    if canonical_device(x.device) != canonical_device(data_device):
         warnings.warn(
             f"Data x has device '{x.device}'. "
             f"Moving x to the data_device '{data_device}'. "
@@ -767,7 +772,7 @@ def validate_theta_and_x(
         )
         x = x.to(data_device)
 
-    if str(theta.device) != str(data_device):
+    if canonical_device(theta.device) != canonical_device(data_device):
         warnings.warn(
             f"Parameters theta has device '{theta.device}'. "
             f"Moving theta to the data_device '{data_device}'. "